+++ battery_notifier/system_tray.py
@@ -2,7 +2,7 @@ import logging
 import time
 
 from battery import BatteryThreshold
-from devices import BaseDevice, Device
+from devices import Device
 from PIL import Image
 from pystray import Icon, Menu, MenuItem, _base
 
@@ -17,6 +17,7 @@ def initialize_system_tray():
         icon=image,
         title="No device found yet",
         menu=Menu(
+            MenuItem(text="Left-Click-Action", action=update_system_tray, default=True, visible=False),
             MenuItem("Quit", Icon.stop),
         ),
     )
@@ -24,12 +25,14 @@ def initialize_system_tray():
 
 
 def battery_loop(system_tray: _base.Icon):
-    all_devices = Device.initialize_all()
     while True:
-        update_system_tray(system_tray, all_devices)
+        update_system_tray(system_tray)
+        logging.info(f"Next check in {CHECK_FREQUENCY} seconds...")
+        time.sleep(CHECK_FREQUENCY)
 
 
-def update_system_tray(system_tray: _base.Icon, all_devices: list[BaseDevice]):
+def update_system_tray(system_tray: _base.Icon):
+    all_devices = Device.initialize_all()
     title_text = ""
     min_battery = 100
     for device in all_devices:
@@ -38,8 +41,7 @@ def update_system_tray(system_tray: _base.Icon, all_devices: list[BaseDevice]):
         if battery_level == -1:
             battery_text = "N/A"
             battery_level = min_battery + 1
-        if battery_level < min_battery:
-            min_battery = battery_level
+        min_battery = min(min_battery, battery_level)
         title_text += f"{device.name}: {battery_text}\n"
 
     title_text = title_text.strip()
@@ -48,5 +50,3 @@ def update_system_tray(system_tray: _base.Icon, all_devices: list[BaseDevice]):
     system_tray.icon = Image.open(battery_threshold.icon())
     if battery_threshold.should_notify():
         system_tray.notify("Low Battery", system_tray.title)
-    logging.info(f"Next check in {CHECK_FREQUENCY} seconds...")
-    time.sleep(CHECK_FREQUENCY)